+++ supabase/migrations/20260704000000_add_auth_verification_to_rpc_functions.sql
@@ -95,14 +95,14 @@ $$;
 -- Verify caller is the driver
 CREATE OR REPLACE FUNCTION withdraw_funds_tx(
   p_driver_id   UUID,
-  p_amount      INT
+  p_amount      numeric
 ) RETURNS void
 LANGUAGE plpgsql
 SECURITY DEFINER
 AS $$
 DECLARE
-  v_confirmed INT;
-  v_pending   INT;
+  v_confirmed numeric;
+  v_pending   numeric;
 BEGIN
   -- Verify the caller IS the driver
   IF auth.uid() <> p_driver_id THEN
